Further fixes and refactors
